Move TemporaryBuffers to be a package protected class, no need to expose this class
