+++ api/trace/src/test/java/io/opentelemetry/api/trace/propagation/TemporaryBuffersTest.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package io.opentelemetry.api.internal;
+package io.opentelemetry.api.trace.propagation;
 
 import static org.assertj.core.api.Assertions.assertThat;
